[refactor]: Controller.js 코드 분리, 일부 오류 해결
1. Controller와 Service가 하는 역할에 맞게 Controller 코드를 분리
- Controller에서는 라우터로 부터 전달 받은 요청을 해석하고 해당 요청에 대한 비즈니스 로직을 수행할 Service를 호출한다.
- Service에서는 Controller로 부터 전달 받은 데이터를 통해 비즈니스 로직을 수행하고 결과를 Controller에게 반환한다.

2. 일부 sql쿼리문 수정

3. 커스텀 에러 생성시, 적절한 위치에서 에러 객체를 throw하지 않아 설정한 에러 메세지가 들어가지 않았던 오류 해결
