+++ src/routes/books.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
 
-const { getBooksInfo, getBookDetail } = require("../controllers/bookController");
+const { getBooksInfo, getBookDetail } = require("../controllers/booksController");
 const { authenticateToken } = require("../middlewares/authMiddleware");
 
 router.get("/", getBooksInfo);
