+++ src/routes/carts.js
@@ -2,7 +2,7 @@ const express = require("express");
 const router = express.Router();
 
 const { authenticateToken } = require("../middlewares/authMiddleware");
-const { addTocart, getCartItems, removeFromCart } = require("../controllers/cartController");
+const { addTocart, getCartItems, removeFromCart } = require("../controllers/cartsController");
 
 router.use(authenticateToken);
