+++ src/routes/categories.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
 
-const { getAllCategories } = require("../controllers/categoryController");
+const { getAllCategories } = require("../controllers/categoriesController");
 
 router.get("/", getAllCategories);
